requirements.txt
# [file name]: requirements.txt
streamlit>=1.50.0
pandas>= 2.3.3
plotly>= 6.3.0
numpy>=2.3.3
langchain>=0.3.27
langchain-openai>=0.3.34
langchain-experimental>=0.3.4
langchain-community>=0.3.30
pinecone-client>=3.1.0
openai>=2.0.1
python-dotenv>=1.1.1